# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FraudWatch API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /analysis:
    get:
      operationId: listAnalyses
      tags: [analysis]
      summary: List analyses
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: fileId
          schema:
            type: string
      responses:
        "200":
          description: List of analyses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisListResponse"
    post:
      operationId: runAnalysis
      tags: [analysis]
      summary: Run fraud analysis on a file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalysisInput"
      responses:
        "201":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analysis"
  /analysis/{id}:
    get:
      operationId: getAnalysis
      tags: [analysis]
      summary: Get analysis details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analysis"
    delete:
      operationId: deleteAnalysis
      tags: [analysis]
      summary: Delete analysis
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ─── Transactions ─────────────────────────────────────────────────────────────
  /analysis/{id}/transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: Get transactions for an analysis
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: prediction
          schema:
            type: string
            enum: [fraud, legitimate, all]
            default: all
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: minRisk
          schema:
            type: number
        - in: query
          name: maxRisk
          schema:
            type: number
      responses:
        "200":
          description: Transaction list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionListResponse"

  # ─── Chat ─────────────────────────────────────────────────────────────────────
  /chat/messages:
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message to AI assistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
  /chat/history/{analysisId}:
    get:
      operationId: getChatHistory
      tags: [chat]
      summary: Get chat history for an analysis
      parameters:
        - in: path
          name: analysisId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Chat history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    delete:
      operationId: clearChatHistory
      tags: [chat]
      summary: Clear chat history for an analysis
      parameters:
        - in: path
          name: analysisId
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Cleared

  # ─── Reports ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    AnalysisInput:
      type: object
      properties:
        fileId:
          type: string
        modelName:
          type: string
          enum: [auto, random_forest, decision_tree, logistic_regression, svm, isolation_forest, xgboost, lightgbm, neural_network]
          default: auto
      required: [fileId]
    ModelMetrics:
      type: object
      properties:
        accuracy:
          type: number
        precision:
          type: number
        recall:
          type: number
        f1Score:
          type: number
        rocAuc:
          type: number
        trainingTimeMs:
          type: integer
        predictionTimeMs:
          type: integer
      required: [accuracy, precision, recall, f1Score, rocAuc, trainingTimeMs, predictionTimeMs]
    RiskBreakdown:
      type: object
      properties:
        critical:
          type: integer
        high:
          type: integer
        medium:
          type: integer
        low:
          type: integer
      required: [critical, high, medium, low]
    Analysis:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        fileId:
          type: string
        fileName:
          type: string
        modelName:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed]
        totalTransactions:
          type: ["integer", "null"]
        fraudCount:
          type: ["integer", "null"]
        legitimateCount:
          type: ["integer", "null"]
        fraudPercentage:
          type: ["number", "null"]
        riskBreakdown:
          $ref: "#/components/schemas/RiskBreakdown"
        metrics:
          $ref: "#/components/schemas/ModelMetrics"
        aiSummary:
          type: ["string", "null"]
        featureImportance:
          type: ["object", "null"]
        errorMessage:
          type: ["string", "null"]
        createdAt:
          type: string
        completedAt:
          type: ["string", "null"]
      required: [id, userId, fileId, fileName, modelName, status, createdAt]
    AnalysisListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Analysis"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [data, total, page, limit, totalPages]

    # ─── Transaction ──────────────────────────────────────────────────────────
    Transaction:
      type: object
      properties:
        id:
          type: string
        analysisId:
          type: string
        transactionId:
          type: string
        amount:
          type: ["number", "null"]
        prediction:
          type: string
          enum: [fraud, legitimate]
        probability:
          type: number
        riskScore:
          type: number
        riskLevel:
          type: string
          enum: [critical, high, medium, low]
        reason:
          type: ["string", "null"]
        rawData:
          type: object
      required: [id, analysisId, transactionId, prediction, probability, riskScore, riskLevel]
    TransactionListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
        fraudCount:
          type: integer
        legitimateCount:
          type: integer
      required: [data, total, page, limit, totalPages, fraudCount, legitimateCount]

    # ─── Chat ────────────────────────────────────────────────────────────────
    ChatMessageInput:
      type: object
      properties:
        message:
          type: string
        analysisId:
          type: ["string", "null"]
      required: [message]
    ChatMessage:
      type: object
      properties:
        id:
          type: string
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        analysisId:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, role, content, createdAt]

    # ─── Report ──────────────────────────────────────────────────────────────
...